* adm_conn.c (kadm_get_ccache): Cast return from getpid() to int 	to match format.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13491 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/kadmin/v5passwdd/adm_conn.c
@@ -138,7 +138,7 @@ kadm_get_ccache(kcontext, user, ccname, ccache, client)
 #ifdef _MACINTOSH
 	(void) sprintf(new_cache, "STDIO:admcc");
 #else
-	(void) sprintf(new_cache, kadm_cache_name_fmt, getpid());
+	(void) sprintf(new_cache, kadm_cache_name_fmt, (int) getpid());
 #endif /* _MACINTOSH */
 #endif /* _MSDOS || _WIN32 */
     }
